--- Develop/.env
@@ -1,3 +1,0 @@
-DB_NAME="chattyCat_db"
-DB_USER=""
-DB_PASSWORD=""
