+++ public/javascript/logout.js
