+++ server.js
@@ -2,7 +2,7 @@ const path =require('path');
 const express = require('express');
 const routes = require('./controllers')
 
-const sequelize = require('./config/connection');
+const sequelize = require('../config/connection');
 
 const app = express();
 const port = process.env.PORT || 3001;
